+++ src/main/java/com/diptanshu/quantummod2/block/custom/GateBlock.java
@@ -13,12 +13,52 @@ import net.minecraft.world.level.block.state.BlockState;
 
 import java.util.Properties;
 
-public class GateBlock extends ButtonBlock {
-    public GateBlock(boolean pressed, BlockBehaviour.Properties properties) {
-        super(pressed, properties);
+import static com.diptanshu.quantummod2.block.custom.QubitBlock.matrixMult;
+
+public abstract class GateBlock extends Block {
+    public GateBlock(BlockBehaviour.Properties properties) {
+        super(properties);
     }
+
+    protected abstract double[] apply(double[] statevector);
 
     protected SoundEvent getSound(boolean p_57062_) {
         return p_57062_ ? SoundEvents.STONE_BUTTON_CLICK_ON : SoundEvents.STONE_BUTTON_CLICK_OFF;
     }
+
+
+    public void onBlockClicked(World worldIn, BlockPos pos, EntityPlayer playerIn){
+        
+        Block blockBelow = level.getBlockState(position.below(1)).getBlock();
+        checkAndUpdate(level, blockBelow, "up");
+
+        Block blockFacingNorth = level.getBlockState(position.north(1)).getBlock();
+        checkAndUpdate(level, blockFacingNorth, "south");
+
+        Block blockFacingSouth = level.getBlockState(position.south(1)).getBlock();
+        checkAndUpdate(level, blockFacingSouth, "north");
+
+        Block blockFacingEast = level.getBlockState(position.east(1)).getBlock();
+        checkAndUpdate(level, blockFacingEast, "west");
+
+        Block blockFacingWest = level.getBlockState(position.west(1)).getBlock();
+        checkAndUpdate(level, blockFacingWest, "east");
+
+        super.onBlockClicked(worldIn, pos,playerIn);
+    }
+
+    public void checkAndUpdate(Level level, Block directionalBlock, String face) {
+        if (level.isClientSide()) {
+            if (directionalBlock instanceof QubitBlock) {
+                QubitBlock qubitBlock = (QubitBlock) directionalBlock;
+                qubitBlock.stateVector = matrixMult(qubitBlock.stateVector, hadamardMatrix);
+            }
+            if (directionalBlock instanceof QubitRegisterBlock) {
+                QubitRegisterBlock registerBlock = (QubitRegisterBlock) directionalBlock;
+                registerBlock.qRegStateVector.replace(face,
+                        matrixMult(registerBlock.qRegStateVector.get(face), hadamardMatrix));
+            }
+        }
+    }
+
 }
