+++ src/main/java/com/diptanshu/quantummod2/block/custom/HadamardBlock.java
@@ -24,36 +24,6 @@ public class HadamardBlock extends GateBlock {
      * @param position
      * @return void
      */
-    public void press(BlockState blockState, Level level, BlockPos position) {
-        Block blockBelow = level.getBlockState(position.below(1)).getBlock();
-        checkAndUpdate(level, blockBelow, "up");
 
-        Block blockFacingNorth = level.getBlockState(position.north(1)).getBlock();
-        checkAndUpdate(level, blockFacingNorth, "south");
 
-        Block blockFacingSouth = level.getBlockState(position.south(1)).getBlock();
-        checkAndUpdate(level, blockFacingSouth, "north");
-
-        Block blockFacingEast = level.getBlockState(position.east(1)).getBlock();
-        checkAndUpdate(level, blockFacingEast, "west");
-
-        Block blockFacingWest = level.getBlockState(position.west(1)).getBlock();
-        checkAndUpdate(level, blockFacingWest, "east");
-
-        super.press(blockState, level, position);
-    }
-
-    public void checkAndUpdate(Level level, Block directionalBlock, String face) {
-        if (level.isClientSide()) {
-            if (directionalBlock instanceof QubitBlock) {
-                QubitBlock qubitBlock = (QubitBlock) directionalBlock;
-                qubitBlock.stateVector = matrixMult(qubitBlock.stateVector, hadamardMatrix);
-            }
-            if (directionalBlock instanceof QubitRegisterBlock) {
-                QubitRegisterBlock registerBlock = (QubitRegisterBlock) directionalBlock;
-                registerBlock.qRegStateVector.replace(face,
-                        matrixMult(registerBlock.qRegStateVector.get(face), hadamardMatrix));
-            }
-        }
-    }
 }
