+++ src/main/java/com/diptanshu/quantummod2/block/custom/QubitBlock.java
@@ -33,14 +33,6 @@ public class QubitBlock extends Block {
 
     public static BlockPos qubitPosition;
 
-    public static double[] matrixMult(double[] vector, double[][] matrix) {
-        double[] resultant = new double[vector.length];
-
-        resultant[0] = round(matrix[0][0] * vector[0] + matrix[0][1] * vector[1], 3);
-        resultant[1] = round(matrix[1][0] * vector[0] + matrix[1][1] * vector[1], 3);
-
-        return resultant;
-    }
 
     @Nullable
     @Override
@@ -52,19 +44,5 @@ public class QubitBlock extends Block {
         stateVector[1] = 0.0;
         printState(pLevel, player, stateVector, "Placement");
         return super.getStateForPlacement(pContext);
-    }
-
-    public static double round(double value, int places) {
-        if (places < 0) throw new IllegalArgumentException();
-
-        BigDecimal number = new BigDecimal(Double.toString(value));
-        number = number.setScale(places, RoundingMode.HALF_UP);
-        return number.doubleValue();
-    }
-
-    public static void printState(Level level, Player player, double[] arr, String label) {
-        if (level.isClientSide()) {
-            player.sendMessage(new TextComponent((label + " state: " + arr[0]) + " |0>  +  " + (arr[1]) + " |1>"), player.getUUID());
-        }
     }
 }
